+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+### Added
+- Support media buttons in editor field (requires WordPress 4.9).
+
 ## [0.3.0]
 ### Added
 - Options page builder